requirements.bak.txt
beautifulsoup4==4.11.2
ctransformers==0.2.27
datasets==2.15.0
diffusers==0.24.0
faster_whisper==0.10.0
fitz==0.0.1.dev2
ipython==8.12.3
langchain==0.3.4
langchain_experimental==0.3.2
pandas==1.5.3
Pillow==10.1.0
progressbar33==2.4
pydantic==2.7.4
pytesseract==0.3.10
python-slugify==8.0.4
pytube==15.0.0
Requests==2.32.3
soundfile==0.12.1
spacy==3.7.2
streamlit==1.29.0
torch==2.2.0
transformers==4.37.0
TTS==0.22.0
huggingface-hub==0.25.0
diffusers

stqdm
pypdf
bitsandbytes
sentence_transformers
faiss-cpu
Pillow
openai
progressbar2
deeplake[enterprise]
tiktoken
IPython
bs4
duckduckgo-search
pytube
pytesseract
numexpr

langchain
langchain-community
streamlit-chat

watchdog
accelerate
sentencepiece
sacremoses
omegaconf
python-slugify
diffusers
invisible_watermark 
safetensors
python-magic
pandas
spacy==3.7.2
faster_whisper
datasets[audio]
torch
torchvision
torchaudio